refactoring code smells

+++ Pinta.Effects/Effects/EdgeDetectEffect.cs
@@ -63,7 +63,7 @@ namespace Pinta.Effects
             }
 
             // adjust and convert angle to radians
-            double r = (double)Data.Angle * 2.0 * Math.PI / 360.0;
+            double r = Data.Angle * 2.0 * Math.PI / 360.0;
 
             // angle delta for each weight
             double dr = Math.PI / 4.0;
